requirements.txt
asgiref==3.7.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-rest-swagger==2.2.0
django-unused-media==0.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pytube3==9.6.4
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.8
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
